+++ src/results-output/results-doc.ts
@@ -190,6 +190,13 @@ export function revealResultsDoc(preserveFocus: boolean = true) {
     });
 }
 
+export async function revealDocForCurrentNS(preserveFocus: boolean = true) {
+    const [_fileName, filePath] = await getFilePathForCurrentNameSpace();
+    vscode.workspace.openTextDocument(filePath).then(doc => vscode.window.showTextDocument(doc, {
+        preserveFocus: false
+    }));
+}
+
 export async function setNamespaceFromCurrentFile() {
     const session = namespace.getSession();
     const ns = namespace.getNamespace(util.getDocument({}));